chore: streamline file structure for upcoming payload cms integration
