+++ next.config.mjs
@@ -1,3 +1,4 @@
+
 import MillionLint from '@million/lint';
 /** @type {import('next').NextConfig} */
 const nextConfig = {};
