+++ src/app/(app)/about/AboutHero.tsx
