+++ src/app/(app)/about/AboutIntro.tsx
