+++ src/app/(app)/about/AboutPartner.tsx
