+++ src/app/(app)/capabilities/CapabilitiesContent.tsx
