+++ src/app/(app)/capabilities/CapabilitiesHero.tsx
