+++ src/app/(app)/capabilities/CapabilitiesWeb.tsx
