+++ src/app/(app)/capabilities/page.tsx
@@ -1,0 +1,14 @@
+import { CapabilitiesHero } from "./CapabilitiesHero";
+import { CapabilitiesBranding } from "./CapabilitiesBranding";
+import { CapabilitiesWeb } from "./CapabilitiesWeb";
+import { CapabilitiesContent } from "./CapabilitiesContent";
+export default function Page() {
+  return (
+    <main>
+      <CapabilitiesHero />
+      <CapabilitiesBranding />
+      <CapabilitiesWeb />
+      <CapabilitiesContent />
+    </main>
+  );
+}
