+++ src/app/(app)/insights/InsightsBody.tsx
