+++ src/app/(app)/insights/InsightsHero.tsx
