+++ src/app/(app)/insights/InsightsWonder.tsx
