+++ src/app/(app)/insights/page.tsx
@@ -1,0 +1,12 @@
+import { InsightsHero } from "./InsightsHero";
+import { InsightsBody } from "./InsightsBody";
+import { InsightsWonder } from "./InsightsWonder";
+export default function Page() {
+  return (
+    <main>
+      <InsightsHero />
+      <InsightsBody />
+      <InsightsWonder />
+    </main>
+  );
+}
