+++ src/app/(app)/layout.tsx
@@ -1,5 +1,5 @@
 import type { Metadata } from "next";
-import { Footer } from "./components/Footer";
+import { Footer } from "../components/Footer";
 import "./globals.css";
 
 export const metadata: Metadata = {
