+++ src/app/(app)/page.tsx
@@ -1,4 +1,4 @@
-import { Landing } from "./components/Landing";
+import { Landing } from "../components/Landing";
 
 export default function Page() {
   return (
